+++ packages/nuomi/src/components/index.js
@@ -9,5 +9,5 @@ export { default as Link } from './Link';
 export { default as NavLink } from './NavLink';
 export { default as connect } from './connect';
 export { default as withNuomi } from './withNuomi';
-export { default as withRouter } from './withRouter';
+export { default as withLocation } from './withLocation';
 export * from './Context';