+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.1.1 (2023-06-19)
 * useNuomi返回值由数组改为对象
-* 增加withRouter高阶组件
+* 增加useLocation hoosk函数，返回location对象
+* 增加withLocation高阶组件，子组件props会增加location对象
 
 ## 1.1.0 (2023-06-17)
 * 取消state中的loading对象，与state数据进行合并，删除reducer中的@loading方法